+++ db/services.php
@@ -3,12 +3,30 @@
 defined('MOODLE_INTERNAL') || die();
 
 $functions = array(
-    'mod_template_getInfo' => array(
-        'classname'   => 'mod_template_external',
+    'mod_hypervideo_getInfo' => array(
+        'classname'   => 'mod_hypervideo_external',
         'methodname'  => 'getInfo',
-        'classpath'   => 'mod/template/db/external.php',
+        'classpath'   => 'mod/hypervideo/db/external.php',
         'description' => 'Get all users to display.',
         'type'        => 'read',
+        'ajax'        => true,
+        'loginrequired' => true
+    ),
+    'mod_hypervideo_log' => array(
+        'classname'   => 'mod_hypervideo_external',
+        'methodname'  => 'log',
+        'classpath'   => 'mod/hypervideo/db/external.php',
+        'description' => 'Get all users to display.',
+        'type'        => 'write',
+        'ajax'        => true,
+        'loginrequired' => true
+    ),
+    'mod_hypervideo_survey' => array(
+        'classname'   => 'mod_hypervideo_external',
+        'methodname'  => 'survey',
+        'classpath'   => 'mod/hypervideo/db/external.php',
+        'description' => 'Get all users to display.',
+        'type'        => 'write',
         'ajax'        => true,
         'loginrequired' => true
     )