+++ view.php
@@ -1,78 +1,66 @@
 <?php
 
-##
-#   Please do not change!
-##
-
 require('../../config.php');
 require_once('lib.php');
 
-$pdata = strtolower(basename(dirname(__DIR__)).'_'.basename(__DIR__)).'\core\pinfo';
-$pdata = new $pdata();
-$pdata = $pdata->getData();
 
-// Change view.php?id=19 to view.php/19/[...]
-$cmid = optional_param('id', 0, PARAM_INT);
-if ($cmid > 0) {
-    $path = "{$pdata->moodlePath}/view.php/{$cmid}/"; 
-    redirect(new \moodle_url($path));
-}
+$id = optional_param('id', 0, PARAM_INT); // Course Module ID
+$g  = optional_param('g', 0, PARAM_INT);  // hypervideo ID
 
-// Get the coursemoduleid
-try{
-    $paths = explode('/', $_SERVER['REQUEST_URI']);
-    $baseindex = array_search('view.php', $paths);
-    if (count($paths) > $baseindex + 1) {
-        $cmid = intval($paths[$baseindex + 1]);
+if (!empty($id)) {
+    if (! $cm = get_coursemodule_from_id('hypervideo', $id)) {
+        print_error('invalidcoursemodule');
     }
-    if(!isset($cmid) || !is_int($cmid) || $cmid <= 0){
-        throw new \Exception(
-            get_string('invcmid', $pdata->fullName)
-        );
-    }    
-    list($course, $coursemodule) = get_course_and_cm_from_cmid($cmid, $pdata->name);
-} catch(Exception $ex){   
-    throw new \Exception(
-        get_string('invcmid', $pdata->fullName)
-    );
+    if (! $course = $DB->get_record("course", array("id"=>$cm->course))) {
+        print_error('coursemisconf');
+    }
+    if (! $hypervideo = $DB->get_record("hypervideo", array("id"=>$cm->instance))) {
+        print_error('invalidid', 'hypervideo');
+    }
+
+} else if (!empty($g)) {
+    if (! $hypervideo = $DB->get_record("hypervideo", array("id"=>$g))) {
+        print_error('invalidid', 'hypervideo');
+    }
+    if (! $course = $DB->get_record("course", array("id"=>$hypervideo->course))) {
+        print_error('invalidcourseid');
+    }
+    if (!$cm = get_coursemodule_from_instance("hypervideo", $hypervideo->id, $course->id)) {
+        print_error('invalidcoursemodule');
+    }
+    $id = $cm->id;
+} else {
+    print_error('invalidid', 'hypervideo');
 }
+require_login($course, true, $cm);
 
-$url = new moodle_url(
-    "{$pdata->moodlePath}/view.php", ['id' => $cmid]
-);
 
-// Require Login
-require_login($course, true, $coursemodule);
-
-$config = get_config($pdata->name);
-
-##
-#   From here on changes can be made!
-##
-
-$permission = $pdata->fullName.'\permission\course';
-$permission = new $permission($USER->id, $course->id);
-
-$title = get_string('modulename', $pdata->fullName);
-
-$PAGE->set_context($coursemodule->context);
-$PAGE->set_title($title);
-$PAGE->set_heading($title);
-$PAGE->set_url($url);
-$PAGE->set_pagelayout('standard');
-
-$PAGE->requires->js_call_amd($pdata->fullName.'/app-lazy', 'init', [
-    'coursemoduleid' => $coursemodule->id,
-    'contextid' => $coursemodule->context->id,
-    'fullPluginName' => $pdata->fullName,
-    'url' => 'http://localhost/movie.mp4',
-    'title' => 'Mein Video'
-]);
+if ($cm) {
+    $PAGE->set_cm($cm, $course); 
+} else {
+    $PAGE->set_course($course);
+}
+$PAGE->set_pagelayout('base');
+$PAGE->set_title($course->shortname.': '.$hypervideo->name);
+$PAGE->set_heading($hypervideo->name);
+$PAGE->set_url('/mod/hypervideo/view.php', array('id' => $cm->id));
 
 echo $OUTPUT->header();
-
+//echo '';
 echo <<<'EOT'
 <div id="app"></div>
 EOT;
-
+$PAGE->requires->js_call_amd('mod_hypervideo/app-lazy', 'init', [
+    'coursemoduleid' => $cm->id,
+    'contextid' => 0,//$cm->context->id,
+    'fullPluginName' => 'mod_hyperaudio',
+    'url' => $hypervideo->url,
+    'title' => 'Video: '. $hypervideo->name
+]);
 echo $OUTPUT->footer($course);
+
+//$json = '{"a":1,"b":2,"c":3,"d":4,"e":5}';
+//var_dump(json_decode($json));
+//var_dump(unserialize($json));
+
+